Try to simplify code
